requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio
ultralytics

flask==3.0.2
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.4
mysql-connector-python==8.3.0 